backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
scikit-image==0.22.0
numpy>=1.26.0
pillow>=10.0.0
python-multipart==0.0.6
python-jose==3.3.0
requests>=2.31.0
